backend/requirements.txt
Django==3.2.13
django-celery-beat==2.2.1
django-cors-headers==3.12.0
django-environ==0.8.1
django-filter==2.4.0
django-timezone-field==4.2.3
djangorestframework==3.12.4
python-dateutil==2.8.2
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
requests==2.27.1
